+++ NEWS.md
@@ -1,5 +1,7 @@
 * update data
 * fixed plot 2020 03 18 intensive veneto
+* 20200325 modul veneto hosp
+* fix colour focus_picco (#24)
 
 # covid19ita 0.19.0