+++ NEWS.md
@@ -1,3 +1,5 @@
+# covid19ita 0.20.0
+
 * update data
 * fixed plot 2020 03 18 intensive veneto
 * 20200325 modul veneto hosp